pyproject.toml
[project]
name = "reflection-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "cal-dortiz",email = "djortiz@berkeley.edu"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.0,<7.0.0)",
    "langchain (>=0.3.17,<0.4.0)",
    "langchain-openai (>=0.3.3,<0.4.0)",
    "langgraph (>=0.2.69,<0.3.0)",
    "langchain-community (>=0.3.16,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
